Move Arcade implementation to live_component
